Move testing Glue metastore creation to test code
